add test and refactor
