+++ internal/bootcamp/encode/encode_test.go
@@ -1,4 +1,4 @@
-package transport
+package encode
 
 import (
 	"github.com/stretchr/testify/assert"
