+++ internal/bootcamp/encode/mocks/EncodeResponse.go
